Update for Markov chain Monte Carlo

This function computes a single update for a Metropolis-Hastings algorithm for univariate proposal based on random walk, adjusted Langevin dynamics with gradient or a quadratic Laplace approximation.

mh(
  type = c("rw", "mala", "laplace"),
  par_curr,
  par_name,
  loglik,
  loglik_grad = NULL,
  loglik_hessian = NULL,
  logprior,
  logprior_grad = NULL,
  logprior_hessian = NULL,
  lb = -Inf,
  ub = Inf,
  transform = FALSE,
  damping = 1,
  sd_prop,
  ...
)

Arguments

type

[string] update type, either random walk (rw), adjusted Langevin (mala) or quadratic approximation (laplace)

par_curr

[double] current value of the parameter

par_name

[string] the name of the argument for the function

loglik

[function] log likelihood function

loglik_grad

[function] derivative of the log likelihood with respect to parameter of interest

loglik_hessian

[function] second derivative of the log likelihood with respect to the parameter of interest

logprior

[function] log prior of parameter

logprior_grad

[function] derivative of the log prior with respect to parameter of interest

logprior_hessian

[function] second derivative of the log prior with respect to the parameter of interest

lb

[scalar] lower bound for the parameter

ub

[scalar] upper bound for the parameter

transform

[logical] should the parameter be mapped onto the real line (TRUE) or should proposals be made based on truncated distributions (FALSE)

damping

[scalar] contraction factor for the Newton update

sd_prop

[double] standard deviation of the proposal

...

additional arguments passed to the log likelihood function and its derivatives

Value

value of the parameter