Electoral participation

Data from a Ipsos survey commissioned by FiveThirtyEight to study determinant of electoral participation during national elections. Data have been coupled with census registers and weighted to reflect the US population as a whole. For more details, consult the description and the survey questionaire. Data were used in the article Why Millions of Americans Don't Vote.

voting

Format

A data frame with 5836 rows and 8 variables:

age

[integer] age of respondant

race

[factor] race, one of black, hispanic, other/mixed or white

gender

[factor] gender, either male or female

income

[factor] income category

vote

[ordered] frequency of vote in national election, either rarely/never, sporadic or always

educ

[factor] education level of respondent

weight

[double] sampling weight

party

[factor] party affiliation, either Democrat (Dem), Republican (GOP) or other for non-response, independent, third party or no preference

Source

FiveThirtyEight, `non-voters` data

Note

Data shared under Creative Commons Attribution 4.0 International License